merged

+++ database/migrations/2025_01_23_135516_remove_test_column_from_career_table.php
@@ -14,7 +14,9 @@ class RemoveTestColumnFromCareerTable extends Migration
     public function up()
     {
         Schema::table('career', function (Blueprint $table) {
-            $table->dropColumn('test');
+            if (Schema::hasColumn('career', 'test')) {
+                $table->dropColumn('test');
+            }
         });
     }
